Revert "Upgrade kURL to Kubernetes 1.16.4"
