+++ addons/nodeless/README.md
@@ -7,7 +7,7 @@ Enable the aws, nodeless and calico plugins, and disable prometheus, weave, rook
     --- a/scripts/Manifest
     +++ b/scripts/Manifest
     @@ -1,12 +1,15 @@
-     KUBERNETES_VERSION=1.16.4
+     KUBERNETES_VERSION=1.15.3
      DOCKER_VERSION=18.09.8
     -WEAVE_VERSION=2.5.2
     -ROOK_VERSION=1.0.4
