+++ deploy/kurl-util/Dockerfile
@@ -6,7 +6,7 @@ COPY go.sum go.sum
 COPY cmd cmd
 
 RUN go build -o proxy cmd/proxy/main.go
-RUN go build -o yamlutil cmd/yamlutil/main.go
+RUN go build -o removefield cmd/removefield/main.go
 
 WORKDIR /go/src/github.com/replicatedhq/kurl/cmd/join
 RUN go build -o join main.go
@@ -16,7 +16,7 @@ FROM debian:buster-slim
 COPY --from=0 /go/src/github.com/replicatedhq/kurl/proxy /usr/local/bin
 COPY cmd/proxy/assets /assets
 
-COPY --from=0 /go/src/github.com/replicatedhq/kurl/yamlutil /usr/local/bin
+COPY --from=0 /go/src/github.com/replicatedhq/kurl/removefield /usr/local/bin
 
 COPY --from=0 /go/src/github.com/replicatedhq/kurl/cmd/join/join /usr/local/bin
