+++ scripts/Manifest
@@ -1,4 +1,4 @@
-KUBERNETES_VERSION=1.16.4
+KUBERNETES_VERSION=1.15.3
 DOCKER_VERSION=18.09.8
 WEAVE_VERSION=2.5.2
 ROOK_VERSION=1.0.4
