+++ scripts/common/rook.sh
@@ -19,3 +19,4 @@ function enable_rook_ceph_operator() {
 function is_rook_1() {
     kubectl -n rook-ceph get cephblockpools replicapool &>/dev/null
 }
+
